refactor: split handlers & data
