+++ main.go
@@ -3,8 +3,7 @@ package main
 import (
 	"net/http"
 
-	"go-service-v2/internal/albums"
-
+	"github.com/cahllagerfeld/go-service-v2/internal/albums/handlers"
 	"github.com/gorilla/mux"
 )
 
@@ -12,6 +11,6 @@ func main() {
 	r := mux.NewRouter()
 	getRouter := r.Methods("GET").Subrouter()
 
-	getRouter.HandleFunc("/", albums.ListAlbums)
+	getRouter.HandleFunc("/", handlers.ListAlbums)
 	http.ListenAndServe(":9090", r)
 }
